pyproject.toml
[project]
name = "StudyRepo"
version = "0.1.0"
description = ""
authors = [
    {name = "vladislav1029"}
]
readme = "README.md"
requires-python = ">=3.11, <4"
dependencies = [
    "django (==4.2.24)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "django-ninja (>=1.4.3,<2.0.0)",
    "django-extensions (>=4.1,<5.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false